+++ src/ui/camera.ts
@@ -1638,6 +1638,6 @@ export abstract class Camera extends Evented {
         if (!this.terrain) {
             return null;
         }
-        return this.terrain.getElevationForLngLat(LngLat.convert(lngLatLike));
+        return this.terrain.getElevationForLngLat(LngLat.convert(lngLatLike), this.transform);
     }
 }